% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-render.R
\name{flex_coords}
\alias{flex_coords}
\title{Create a data.frame of text coordiantes for the given text}
\usage{
flex_coords(s, w = 1, h = 1, hgap = 0.1, vgap = 0.1, npoints = 10)
}
\arguments{
\item{s}{string. May include newlines}

\item{w, h}{character width and height}

\item{hgap, vgap}{the gap between each letter}

\item{npoints}{the number of points to render for each arc.  Default: 10.
Higher numbers will give a smoother appearance to the curves.}
}
\value{
data.frame of coordinates
}
\description{
Create a data.frame of text coordiantes for the given text
}
\examples{
flex_coords('x')
}
