% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-render.R
\name{flextextGrob}
\alias{flextextGrob}
\title{Create a grob to represetnt the given string in flex font}
\usage{
flextextGrob(
  txt,
  w = 1,
  h = 1,
  scale = 1,
  hgap = 0.1,
  vgap = 0.1,
  x = unit(0.5, "npc"),
  y = unit(0.5, "npc"),
  hjust = 0.5,
  vjust = 0.5,
  default.units = "npc",
  gp = grid::gpar(),
  npoints = 10
)
}
\arguments{
\item{txt}{chracter string. May contain newlines. Only letters A-Z are
rendered}

\item{w, h}{the size of each letter}

\item{scale}{A scaling factor applied to the letter size. Default: 1}

\item{hgap, vgap}{the gap between each letter}

\item{x, y}{location of text on page. Default: centre}

\item{hjust, vjust}{justificaiton for this block of text. Default 0.5 (middle)}

\item{default.units}{default units for grid rendering. Default: 'npc'}

\item{gp}{graphics parameters e.g. lwd, color, linetype}

\item{npoints}{the number of points to render for each arc.  Default: 10.
Higher numbers will give a smoother appearance to the curves.}
}
\value{
grid graphics object
}
\description{
Create a grob to represetnt the given string in flex font
}
\examples{
g <- flextextGrob("hello")
grid::grid.draw(g)
}
